refactor(tests): rename folder and namespace for RoleTestHelper
