+++ tests/Traits/RoleTestHelper.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace Tests\Helpers;
+namespace Tests\Traits;
 
 use Spatie\Permission\Models\Role;
 use Spatie\Permission\Models\Permission;
